refactor (StatisticsLottoDto.java) : 분석 결과를 전달 하는 DTO 분리
